Move methods from ModifierCorrectionSubProcessor into core class.
- Make ModifierCorrectionSubProcessor subclass the abstract class,
  ModifierCorrectionSubProcessorCore and move methods into the base
